requirements.txt
chainlit==0.2.1
langchain==0.0.10
transformers==4.25.1
torch==1.12.1 
pinecone-client==2.0.0
PyPDF2==1.26.0
requests==2.28.1
sentence-transformers==2.2.0
